refactor of components
